+++ src/lib/components/interactables/button/Button.svelte
@@ -4,6 +4,7 @@
     import type {PROPERTY_PALETTE} from "../../../types/palettes";
     import type {PROPERTY_SIZING} from "../../../types/sizings";
     import type {IMarginProperties} from "../../../types/spacings";
+    import type {ISvelteKitAnchorProperties} from "../../../types/sveltekit";
     import type {PROPERTY_VARIATION_BUTTON} from "../../../types/variations";
 
     import {behavior_button} from "../../../actions/behavior_button";
@@ -37,11 +38,16 @@
         for?: string;
 
         palette?: PROPERTY_PALETTE;
+        /**
+         * @deprecated Use `<Button sizing="...">` instead.
+         */
         size?: PROPERTY_SIZING;
+        sizing?: PROPERTY_SIZING;
         variation?: PROPERTY_VARIATION_BUTTON;
     } & IHTML5Properties &
         IGlobalProperties &
-        IMarginProperties;
+        IMarginProperties &
+        ISvelteKitAnchorProperties;
 
     type $$Slots = {
         default: {};
@@ -69,7 +75,11 @@
     export {_for as for};
 
     export let palette: $$Props["palette"] = undefined;
+    /**
+     * @deprecated Use `<Button sizing="...">` instead.
+     */
     export let size: $$Props["size"] = undefined;
+    export let sizing: $$Props["sizing"] = undefined;
     export let variation: $$Props["variation"] = undefined;
 
     // HACK: Svelte has `tabindex` typed as `number | undefined` unless
@@ -83,7 +93,7 @@
         {...map_global_attributes($$props)}
         role="button"
         class="button {_class}"
-        {...map_data_attributes({palette, size, variation})}
+        {...map_data_attributes({palette, size: size ?? sizing, variation})}
         {...map_aria_attributes({active, disabled})}
         {download}
         {href}
@@ -115,7 +125,7 @@
         class="button {_class}"
         for={_for}
         tabindex={_tabindex}
-        {...map_data_attributes({palette, size, variation})}
+        {...map_data_attributes({palette, size: size ?? sizing, variation})}
         {...map_aria_attributes({disabled, pressed: active})}
         use:behavior_button={{enabled: true}}
         use:forward_actions={{actions}}
@@ -142,7 +152,7 @@
             bind:this={element}
             {...map_global_attributes($$props)}
             type="reset"
-            {...map_data_attributes({palette, size, variation})}
+            {...map_data_attributes({palette, size: size ?? sizing, variation})}
             {...map_aria_attributes({pressed: active})}
             {...map_attributes({disabled, value})}
             use:forward_actions={{actions}}
@@ -166,7 +176,7 @@
             bind:this={element}
             {...map_global_attributes($$props)}
             type="submit"
-            {...map_data_attributes({palette, size, variation})}
+            {...map_data_attributes({palette, size: size ?? sizing, variation})}
             {...map_aria_attributes({pressed: active})}
             {...map_attributes({disabled, value})}
             use:forward_actions={{actions}}
@@ -190,7 +200,7 @@
             bind:this={element}
             {...map_global_attributes($$props)}
             type="button"
-            {...map_data_attributes({palette, size, variation})}
+            {...map_data_attributes({palette, size: size ?? sizing, variation})}
             {...map_aria_attributes({pressed: active})}
             {...map_attributes({disabled, value})}
             use:forward_actions={{actions}}
@@ -214,7 +224,7 @@
     <button
         bind:this={element}
         {...map_global_attributes($$props)}
-        {...map_data_attributes({palette, size, variation})}
+        {...map_data_attributes({palette, size: size ?? sizing, variation})}
         {...map_aria_attributes({pressed: active})}
         {...map_attributes({disabled})}
         use:forward_actions={{actions}}