+++ src/lib/components/layouts/mosaic/index.ts
@@ -1,1 +1,6 @@
-export {default as Mosaic} from "./Mosaic.svelte";
+import _Mosaic from "./Mosaic.svelte";
+
+/**
+ * @deprecated Updated into a multi-part `<Mosaic.Container>` Component in `0.6.0`.
+ */
+export const Mosaic = _Mosaic;