+++ src/lib/components/overlays/overlay/OverlayButton.svelte
@@ -21,8 +21,12 @@
         active?: boolean;
         disabled?: boolean;
 
-        palette?: PROPERTY_PALETTE;
+        palette?: PROPERTY_PALETTE
+        /**
+         * @deprecated Use `<Overlay.Button sizing="...">` instead.
+         */;
         size?: PROPERTY_SIZING;
+        sizing?: PROPERTY_SIZING;
         variation?: PROPERTY_VARIATION_BUTTON;
     } & IHTML5Properties &
         IGlobalProperties &
@@ -41,7 +45,11 @@
     export let disabled: $$Props["disabled"] = undefined;
 
     export let palette: $$Props["palette"] = undefined;
+    /**
+     * @deprecated Use `<Overlay.Button sizing="...">` instead.
+     */
     export let size: $$Props["size"] = undefined;
+    export let sizing: $$Props["sizing"] = undefined;
     export let variation: $$Props["variation"] = undefined;
 
     const _overlay_id = CONTEXT_OVERLAY_ID.get();
@@ -57,11 +65,11 @@
     {...$$props}
     bind:element
     for={$_overlay_id}
+    size={size ?? sizing}
     {actions}
     {active}
     {disabled}
     {palette}
-    {size}
     {variation}
     {tabindex}
     on:click