+++ src/lib/components/typography/heading/Heading.svelte
@@ -23,7 +23,11 @@
 
         is?: "h1" | "h2" | "h3" | "h4" | "h5" | "h6";
 
+        /**
+         * @deprecated Use `<Heading alignment_x="...">` instead.
+         */
         align?: PROPERTY_TEXT_ALIGNMENT;
+        alignment_x?: PROPERTY_TEXT_ALIGNMENT;
         transform?: PROPERTY_TEXT_TRANSFORM;
 
         palette?: PROPERTY_PALETTE;
@@ -46,7 +50,11 @@
 
     export let is: $$Props["is"] = "h1";
 
+    /**
+     * @deprecated Use `<Heading alignment_x="...">` instead.
+     */
     export let align: $$Props["align"] = undefined;
+    export let alignment_x: $$Props["alignment_x"] = undefined;
     export let transform: $$Props["transform"] = undefined;
 
     export let palette: $$Props["palette"] = undefined;
@@ -58,7 +66,7 @@
         bind:this={element}
         {...map_global_attributes($$props)}
         class="heading {_class}"
-        {...map_data_attributes({align, palette, transform, variation})}
+        {...map_data_attributes({align: align ?? alignment_x, palette, transform, variation})}
         use:forward_actions={{actions}}
         on:click
         on:contextmenu
@@ -82,7 +90,7 @@
         bind:this={element}
         {...map_global_attributes($$props)}
         class="heading {_class}"
-        {...map_data_attributes({align, palette, transform, variation})}
+        {...map_data_attributes({align: align ?? alignment_x, palette, transform, variation})}
         use:forward_actions={{actions}}
         on:click
         on:contextmenu
@@ -106,7 +114,7 @@
         bind:this={element}
         {...map_global_attributes($$props)}
         class="heading {_class}"
-        {...map_data_attributes({align, palette, transform, variation})}
+        {...map_data_attributes({align: align ?? alignment_x, palette, transform, variation})}
         use:forward_actions={{actions}}
         on:click
         on:contextmenu
@@ -130,7 +138,7 @@
         bind:this={element}
         {...map_global_attributes($$props)}
         class="heading {_class}"
-        {...map_data_attributes({align, palette, transform, variation})}
+        {...map_data_attributes({align: align ?? alignment_x, palette, transform, variation})}
         use:forward_actions={{actions}}
         on:click
         on:contextmenu
@@ -154,7 +162,7 @@
         bind:this={element}
         {...map_global_attributes($$props)}
         class="heading {_class}"
-        {...map_data_attributes({align, palette, transform, variation})}
+        {...map_data_attributes({align: align ?? alignment_x, palette, transform, variation})}
         use:forward_actions={{actions}}
         on:click
         on:contextmenu
@@ -178,7 +186,7 @@
         bind:this={element}
         {...map_global_attributes($$props)}
         class="heading {_class}"
-        {...map_data_attributes({align, palette, transform, variation})}
+        {...map_data_attributes({align: align ?? alignment_x, palette, transform, variation})}
         use:forward_actions={{actions}}
         on:click
         on:contextmenu