+++ src/lib/types/sveltekit.ts
@@ -1,0 +1,4 @@
+export interface ISvelteKitAnchorProperties {
+    "sveltekit:prefetch"?: boolean;
+    "sveltekit:noscroll"?: boolean;
+}